refactor: Move player retrivial function from BaseAudioCommandModule to MusicBotHelper
